Create the yabf version of configfile modules list
