+++ src/main.js
@@ -2,6 +2,7 @@ export * from './core/term-application'
 
 export * from './commands/scripts-list.command'
 export * from './commands/scripts-run.command'
+export * from './commands/modules-list.command'
 
 export * from './services/file.service'
 export * from './services/exec'
